backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.24.1
openai==0.27.8
python-dotenv==1.0.0
